+++ model_arena/model_arena/__init__.py
@@ -1,7 +1,4 @@
-from . import core
-from . import extensions
-from . import modules
 from .arena import *
 
 # define version
-__version__ = "0.0.1"
+__version__ = "0.0.2dev"