+++ model_arena/CHANGELOG.md
@@ -1,6 +1,12 @@
 # CHANGELOG
 
-## v0.0.1 - 2024-MM-DD
+## v0.0.2 - 2024-MM-DD
+
+### Refactored
+
+- Refact extensions to have higher compatibility. ([#10](https://code.byted.org/smart-infra/ModelArena/merge_requests/10))
+
+## v0.0.1 - 2024-04-17
 
 ### Added
 
@@ -10,6 +16,7 @@
 - All get methods support multiple models and datasets. ([#5](https://bits.bytedance.net/code/smart-infra/ModelArena/merge_requests/5))
 - Add core llms module. ([#6](https://bits.bytedance.net/code/smart-infra/ModelArena/merge_requests/6))
 - Add needle in a haystack analyst. ([#7](https://bits.bytedance.net/code/smart-infra/ModelArena/merge_requests/7))
+- Add gpt as a pairwise evaluator. ([#8](https://bits.bytedance.net/code/smart-infra/ModelArena/merge_requests/8))
 
 ### Refactored